organizing project
- removing loginViewModel and RegisterViewModel
