+++ BancoXYZ/Views/LoginView.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="BancoXYZ.Views.RegisterView"
+﻿<Window x:Class="BancoXYZ.Views.LoginView"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -14,7 +14,7 @@
 
     <Border CornerRadius="20">
         <Border.Background>
-            <LinearGradientBrush StartPoint="1,1" EndPoint="0,0">
+            <LinearGradientBrush StartPoint="0,0" EndPoint="1,1">
                 <GradientStop Color="White" Offset="0.0"/>
                 <GradientStop Color="White" Offset="0.5"/>
                 <GradientStop Color="#0170a1" Offset="0.8"/>
@@ -25,14 +25,15 @@
             <Button Content="_" HorizontalAlignment="Right" VerticalAlignment="Top" Width="30" Height="30" Click="MinimizeButton_Click" Margin="0,0,53,0" Background="Transparent"/>
             <Button Content="X" HorizontalAlignment="Right" VerticalAlignment="Top" Width="30" Height="30" Margin="0,0,21,0" Click="CloseButton_Click" Background="Transparent"/>
             <StackPanel HorizontalAlignment="Center" VerticalAlignment="Center">
+                <TextBlock x:Name="BankNamePlaceholder" Margin="0,10,0,0" Text="XYZ Bank" Foreground="#282735" FontSize="24" FontWeight="Bold" HorizontalAlignment="Center"/>
+                <TextBlock x:Name="AgencyPlaceholder" Width="200" Margin="0,10,0,0" Text="Agency" Foreground="Gray"/>
+                <TextBox x:Name="AgencyTextBox" Text="01" IsEnabled="false" Width="200" Margin="0,10,0,0" Style="{StaticResource CustomTextBoxStyle}"/>
                 <TextBlock x:Name="CurrentAccountPlaceholder" Width="200" Margin="0,10,0,0" Text="Current Account" Foreground="Gray"/>
-                <TextBox x:Name="CurrentAccountTextBox" Width="200" Margin="0,10,0,0" PreviewTextInput="CurrentAccountTextBox_PreviewTextInput" Style="{StaticResource CustomTextBoxStyle}"/>
-                <TextBlock x:Name="NamePlaceholder" Width="200" Margin="0,10,0,0" Text="Name" Foreground="Gray"/>
-                <TextBox x:Name="NameTextBox" Width="200" Margin="0,10,0,0" Style="{StaticResource CustomTextBoxStyle}"/>
+                <TextBox x:Name="CurrentAccountTextBox" Width="200" Margin="0,10,0,0" Style="{StaticResource CustomTextBoxStyle}"/>
                 <TextBlock x:Name="PasswordPlaceholder" Text="Password" Foreground="Gray" Margin="0,10,0,0" IsHitTestVisible="False" VerticalAlignment="Center" HorizontalAlignment="Left" Visibility="Visible"/>
                 <PasswordBox x:Name="PasswordTextBox" Width="200" Margin="0,10,0,0" Style="{StaticResource CustomPasswordBoxStyle}"/>
-                <Button Content="Register" Width="200" Margin="0,30,0,10" Click="OnRegisterClick" Style="{StaticResource CustomButtonStyle}"/>
-                <Button Content="Back to Login" Width="200" Margin="0,10,0,0" Click="OnBackToLoginClick" Style="{StaticResource CustomButtonStyle}"/>
+                <Button Content="Enter" Width="200" Margin="0,30,0,10" Click="OnLoginClick" Style="{StaticResource CustomButtonStyle}"/>
+                <Button Content="New Account" Width="200" Margin="0,10,0,0" Click="OnRegisterClick" Style="{StaticResource CustomButtonStyle}"/>
             </StackPanel>
         </Grid>
     </Border>
