--- BancoXYZ/Views/MainView.xaml
@@ -1,33 +1,0 @@
-﻿<Window x:Class="BancoXYZ.Views.MainView"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:BancoXYZ.Views"
-        mc:Ignorable="d"
-        Height="600" Width="920"
-        WindowStyle="None"
-        ResizeMode="NoResize"
-        Background="Transparent"
-        AllowsTransparency="True"
-        WindowStartupLocation="CenterScreen">
-
-    <Border Background="#272537"
-            CornerRadius="20">
-        <StackPanel HorizontalAlignment="Center" VerticalAlignment="Center">
-            <TextBlock HorizontalAlignment="Center" Margin="0,20,0,20">
-            <Run Text="XYZ" FontSize="36" FontWeight="Bold" Foreground="White"/>
-            <Run Text=" Bank" FontSize="24" FontWeight="Normal" Foreground="White"/>
-            </TextBlock>
-
-            <TextBlock x:Name="AgencyPlaceholder" Width="200" Margin="10,0,10,0" Text="Agency" Foreground="Gray"/>
-            <TextBox x:Name="AgencyTextBox" Width="200" Text="01" IsEnabled="False" Margin="0,10,0,0" Style="{StaticResource DisabledTextBoxStyle}"/>
-            <TextBlock x:Name="CurrentAccountPlaceholder" Width="200" Margin="0,10,0,0" Text="Current Account" Foreground="Gray"/>
-            <TextBox x:Name="CurrentAccountTextBox" Width="200" Margin="0,10,0,0" PreviewTextInput="CurrentAccountTextBox_PreviewTextInput" Style="{StaticResource CustomTextBoxStyle}"/>
-            <TextBlock x:Name="PasswordPlaceholder" Text="Password" Foreground="Gray" Margin="10,10,0,0" IsHitTestVisible="False" VerticalAlignment="Center" HorizontalAlignment="Left" Visibility="Visible"/>
-            <PasswordBox x:Name="PasswordTextBox" Width="200" Margin="0,10,0,0" Style="{StaticResource CustomPasswordBoxStyle}"/>
-            <Button Content="Enter" Width="200" Margin="0,30,0,10" Click="OnLoginClick" Style="{StaticResource CustomButtonStyle}"/>
-            <Button Content="New Account" Width="200" Margin="0,10,0,0" Click="OnRegisterClick" Style="{StaticResource CustomButtonStyle}"/>
-        </StackPanel>
-    </Border>
-</Window>
